divice executor to new file
